ejercicios PY 05/11/25

+++ PYTHON/Ejercicios basicos/ejercicio3.5.py
@@ -1,4 +1,4 @@
-num1=int(input("dame el primer numero: "))
-num2=int(input("dame el segundo numero: "))
+num1=float(input("dame el primer numero: "))
+num2=float(input("dame el segundo numero: "))
 media=(num1+num2)/2
 print("la media de tus numeros es " ,media)
